% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfort.R
\name{ck_humidex}
\alias{ck_humidex}
\title{Humidex}
\usage{
ck_humidex(tavg, dewpoint)
}
\arguments{
\item{tavg}{Numeric vector of temperatures (degrees C).}

\item{dewpoint}{Numeric vector of dewpoint temperatures (degrees C).}
}
\value{
A data frame with columns \code{value}, \code{index}, and \code{unit}.
}
\description{
Compute the Canadian humidex from temperature and dewpoint.
}
\examples{
ck_humidex(tavg = c(30, 35), dewpoint = c(20, 25))
}
\references{
Masterson, J., & Richardson, F. A. (1979). Humidex: A method
of quantifying human discomfort due to excessive heat and humidity.
Environment Canada.
}
