% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear Cache}
\usage{
clear_cache()
}
\value{
Invisibly returns \code{TRUE} if cache was cleared, \code{FALSE} if no cache
existed.
}
\description{
Removes any cached reference data stored by \code{climatekit}.
}
\examples{
\donttest{
op <- options(climatekit.cache_dir = tempdir())
clear_cache()
options(op)
}
}
